+++ include/kernel/infrastructure/i686/drivers/iodelay.h
@@ -29,8 +29,8 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef JINUE_KERNEL_INFRASTRUCTURE_I686_IODELAY_H
-#define JINUE_KERNEL_INFRASTRUCTURE_I686_IODELAY_H
+#ifndef JINUE_KERNEL_INFRASTRUCTURE_I686_DRIVERS_IODELAY_H
+#define JINUE_KERNEL_INFRASTRUCTURE_I686_DRIVERS_IODELAY_H
 
 void iodelay(void);
