+++ kernel/Makefile
@@ -122,12 +122,12 @@ sources.kernel.c = \
 sources.kernel.nasm = \
 	infrastructure/i686/atomic/atomic.asm \
 	infrastructure/i686/atomic/spinlock.asm \
+	infrastructure/i686/drivers/iodelay.asm \
 	infrastructure/i686/isa/abi.asm \
 	infrastructure/i686/isa/instrs.asm \
 	infrastructure/i686/isa/io.asm \
 	infrastructure/i686/isa/regs.asm \
 	infrastructure/i686/pmap/init.asm \
-	infrastructure/i686/iodelay.asm \
 	infrastructure/i686/thread.asm \
 	interface/i686/crt.asm \
 	interface/i686/trap.asm
